Add comments and new tests
